fix: resolve SwiftLint violations in URL scheme and settings code
Extract HooksSettingsTab to its own file (SettingsWindow.swift exceeded
600-line limit), break up cyclomatic complexity in URLRoute.parse and
URLRoute.execute by extracting helper methods, add findSource(named:)
to StatusService to DRY case-insensitive lookups, and fix a 249-char
line length violation.
